Minor updates to backends/rendering

+++ src/consts.hpp
@@ -1,5 +1,5 @@
 #pragma once
 
 const double G = 6.67408e-11 * 1e10;
-const double Geps = 1e-1;
+const float Geps = 1e-1;
